refactor: PAGE ROUTE CHANGE
