+++ app/archive/project/[slug]/page.tsx
@@ -1,7 +1,7 @@
 import { Metadata } from 'next';
 import { allProjects } from '@contentlayer/generated';
 
-import ProjectArticle from '../../../features/project/ProjectArticle';
+import ProjectArticle from '../../../../features/project/ProjectArticle';
 
 export const metadata: Metadata = {
   title: 'ISA Archive - Project Article',
