+++ app/work/page.tsx
@@ -1,0 +1,3 @@
+export default function WorkPage() {
+return <main><h1>work</h1></main>;
+}
